requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
